+++ GloveForDeaf.X/glove.h
@@ -12,15 +12,15 @@
 
 typedef struct
 {
-  uint8_t SIZE;
+  int SIZE;
   flex_t* FLEX;
 }glove_t;
 
-void GLOVE_INIT(glove_t* GLOVE, flex_t FLEX[],uint8_t SIZE);
+void GLOVE_INIT(glove_t* GLOVE, flex_t FLEX[],int SIZE);
 
 void GLOVE_SAMPLE(glove_t* GLOVE);
 
-void GLOVE_GET_STATES(const glove_t*GLOVE, uint8_t STATES_ARR[], uint8_t STATES_ARR_SIZE);
+void GLOVE_GET_STATES(const glove_t*GLOVE, int STATES_ARR[], int STATES_ARR_SIZE);
 
 #endif	/* GLOVE_H */
